+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 <!-- Your comment below here -->
 
 * Add `danger.import_dangerfile(url:)` to import Dangerfile from custom URL - [@dstranz](https://github.com/dstranz)
+* Fixes issue with not being able to update comments, and instead always posting new comments regardless of settings for Bitbucket Cloud [@tskulbru](https://github.com/tskulbru)
 * Fix incorrect command in error message [@revolter](https://github.com/revolter)
 
 <!-- Your comment above here -->